Move dispatcher tests to separate file
